Separate tests into their own module. Add image corruption test
