+++ image_get/worker/worker/dummy_producer.py
