+++ image_get/worker/worker/settings.py
